fix features macro
