+++ crates/dev/orchestrate/Cargo.toml
@@ -17,6 +17,8 @@ poison_girl_macro_def_from_path_buf = { workspace = true }
 strum = { workspace = true }
 strum_macros = { workspace = true }
 toml = { workspace = true }
+poison_girl_dev_util = { workspace = true }
+poison_girl_macro_def_features = { workspace = true }
 
 [dev-dependencies]
 quickcheck = { workspace = true }
