+++ crates/dev/util/src/lib.rs
@@ -1,1246 +1,2 @@
-use {
-	poison_girl_dev_error::{
-		PathIsNotValidUtf8, PathNotFound, PoisonGirlB, ReShape, X, Y,
-		poison_girl_err,
-	},
-	std::path::PathBuf,
-};
-
-pub trait StrEnhanced: CaseConvert + StringKind
-{
-}
-
-pub trait CaseConvert
-{
-	type _Marker;
-	fn is_camel(&self,) -> bool;
-	fn is_snake(&self,) -> bool;
-	fn is_screaming_snake(&self,) -> bool;
-	fn is_kebab(&self,) -> bool;
-
-	fn to_camel<S1: StringKind,>(&self,) -> S1
-	{
-		self.case_transit(
-			|s| format!("{}{}", s[..1].to_ascii_uppercase(), &s[1..]),
-			None,
-		)
-	}
-
-	fn to_snake<S1: StringKind,>(&self,) -> S1
-	{
-		self.case_transit(|s| s.to_ascii_lowercase(), Some('_',),)
-	}
-
-	fn to_screaming_snake<S1: StringKind,>(&self,) -> S1
-	{
-		self.case_transit(|s| s.to_ascii_uppercase(), Some('_',),)
-	}
-
-	fn to_kebab<S1: StringKind,>(&self,) -> S1
-	{
-		self.case_transit(|s| s.to_ascii_lowercase(), Some('-',),)
-	}
-
-	fn case_transit<S: StringKind,>(
-		&self,
-		converter: impl FnMut(String,) -> String,
-		spacer: Option<char,>,
-	) -> S
-	{
-		let converted: Vec<_,> =
-			self.words().into_iter().map(converter,).collect();
-		let spacer = spacer.map_or("".to_string(), |c| c.to_string(),);
-		let converted = converted.join(&spacer,);
-		S::from(converted,)
-	}
-
-	fn find_spacer<S: StringKind,>(&self,) -> Option<S,>;
-	fn words(&self,) -> Vec<String,>;
-	fn as_string_kind(&self,) -> Option<&impl StringKind,>;
-}
-
-pub trait StringKind
-{
-	type DumpReturn;
-	fn dump_string(&self,) -> Self::DumpReturn;
-	fn from(s: impl Into<String,>,) -> Self;
-	fn as_case_convert(&self,) -> Option<&impl CaseConvert,>;
-}
-
-impl StrEnhanced for String
-{
-}
-
-impl CaseConvert for String
-{
-	type _Marker = String;
-
-	fn is_camel(&self,) -> bool
-	{
-		is_xxx_format_with_case(self.clone(), None, Form::StartWithUpper,)
-	}
-
-	fn is_snake(&self,) -> bool
-	{
-		is_xxx_format_with_case(self.clone(), Some('_',), Form::Lower,)
-	}
-
-	fn is_screaming_snake(&self,) -> bool
-	{
-		is_xxx_format_with_case(self.clone(), Some('_',), Form::Upper,)
-	}
-
-	fn is_kebab(&self,) -> bool
-	{
-		is_xxx_format_with_case(self.clone(), Some('-',), Form::Lower,)
-	}
-
-	fn find_spacer<S1: StringKind,>(&self,) -> Option<S1,>
-	{
-		let s: String = self.clone();
-		if s.contains("_",) {
-			Some(S1::from("_".to_string(),),)
-		} else if s.contains("-",) {
-			Some(S1::from("-".to_string(),),)
-		} else {
-			None
-		}
-	}
-
-	fn words(&self,) -> Vec<String,>
-	{
-		let s: String = self.clone();
-		if self.is_camel() {
-			let mut rslt = vec![];
-			let mut idx = 0;
-			while let Some(sub,) = s.get(idx + 1..,)
-				&& let Some(tail,) = sub.find(|c: char| c.is_ascii_uppercase(),)
-			{
-				// tail is relative to sub, so we need to add idx + 1 to get the
-				// absolute position
-				let absolute_pos = idx + 1 + tail;
-				rslt.push(s[idx..absolute_pos].to_string(),);
-				idx = absolute_pos; // Move to the position of the uppercase letter
-			}
-			// Add the remaining part if any
-			if let Some(remaining,) = s.get(idx..,)
-				&& !remaining.is_empty()
-			{
-				rslt.push(remaining.to_string(),);
-			}
-			rslt
-		} else {
-			// Cache the spacer to avoid repeated calls
-			let spacer = s.find_spacer().unwrap_or(" ".to_string(),);
-			s.split(|c: char| spacer == c.to_string(),)
-				.map(|s| s.to_string(),)
-				.collect()
-		}
-	}
-
-	#[allow(refining_impl_trait)]
-	fn as_string_kind(&self,) -> Option<&Self,>
-	{
-		Some(self,)
-	}
-}
-
-impl StringKind for String
-{
-	type DumpReturn = Self;
-
-	fn dump_string(&self,) -> Self::DumpReturn
-	{
-		self.clone()
-	}
-
-	fn from(s: impl Into<String,>,) -> Self
-	{
-		s.into()
-	}
-
-	#[allow(refining_impl_trait)]
-	fn as_case_convert(&self,) -> Option<&Self,>
-	{
-		Some(self,)
-	}
-}
-
-enum Form
-{
-	StartWithUpper,
-	Upper,
-	Lower,
-}
-
-fn is_xxx_format_with_case(
-	s: impl Into<String,> + Clone,
-	spacer: Option<char,>,
-	form: Form,
-) -> bool
-{
-	let s: String = s.into();
-
-	let spacer_checker = || -> Box<dyn Fn(char,) -> bool,> {
-		match spacer {
-			Some(spacer,) => Box::new(move |c| c == spacer,),
-			None => Box::new(|c| c.is_ascii_alphanumeric(),),
-		}
-	};
-	let checker = || -> Box<dyn Fn(&String,) -> bool,> {
-		match form {
-			Form::StartWithUpper => Box::new(|s| {
-				s.starts_with(|c: char| c.is_ascii_uppercase(),)
-					&& s.chars().all(|c| {
-						c.is_ascii_alphanumeric() && spacer_checker()(c,)
-					},)
-			},),
-			Form::Upper => Box::new(|s| {
-				s.chars().all(|c| {
-					c.is_ascii_uppercase()
-						|| c.is_numeric() || spacer_checker()(c,)
-				},)
-			},),
-			Form::Lower => Box::new(|s| {
-				s.chars().all(|c| {
-					c.is_ascii_lowercase()
-						|| c.is_numeric() || spacer_checker()(c,)
-				},)
-			},),
-		}
-	};
-
-	checker()(&s,)
-}
-
-impl StrEnhanced for PathBuf
-{
-}
-
-impl CaseConvert for PathBuf
-{
-	type _Marker = PathBuf;
-
-	fn is_camel(&self,) -> bool
-	{
-		self.dump_string().is_x_and(|s| s.is_camel(),)
-	}
-
-	fn is_snake(&self,) -> bool
-	{
-		self.dump_string().is_x_and(|s| s.is_snake(),)
-	}
-
-	fn is_screaming_snake(&self,) -> bool
-	{
-		self.dump_string().is_x_and(|s| s.is_screaming_snake(),)
-	}
-
-	fn is_kebab(&self,) -> bool
-	{
-		self.dump_string().is_x_and(|s| s.is_kebab(),)
-	}
-
-	fn find_spacer<S: StringKind,>(&self,) -> Option<S,>
-	{
-		// let spacer = self.dump_string().map(|s| s.find_spacer(),)?;
-		// Some(spacer,)
-		match self.dump_string() {
-			X(s,) => s.find_spacer(),
-			Y(e,) => {
-				// to avoid panic! macro
-				eprintln!("{e}");
-				None
-			},
-		}
-	}
-
-	fn words(&self,) -> Vec<String,>
-	{
-		match self.dump_string() {
-			X(s,) => s.words(),
-			Y(e,) => {
-				eprintln!("{e}");
-				vec![]
-			},
-		}
-	}
-
-	#[allow(refining_impl_trait)]
-	fn as_string_kind(&self,) -> Option<&Self,>
-	{
-		Some(self,)
-	}
-}
-
-impl StringKind for PathBuf
-{
-	type DumpReturn = PoisonGirlB<String,>;
-
-	fn dump_string(&self,) -> Self::DumpReturn
-	{
-		self.file_prefix()
-			.reshape(poison_girl_err!(PathNotFound(
-				"failed to get file/dir name".to_string()
-			)),)?
-			.to_str()
-			.reshape(poison_girl_err!(PathIsNotValidUtf8),)
-			.map(|s| s.to_string(),)
-	}
-
-	#[cold]
-	#[track_caller]
-	#[expect(
-		clippy::unreachable,
-		reason = "indicates method that should not use but required to \
-		          implement trait"
-	)]
-	fn from(_: impl Into<String,>,) -> Self
-	{
-		unreachable!("you should not use `PathBuf::from`")
-		// let s: String = s.into();
-		// let s = s.as_str();
-		// PathBuf::from_str(s,).expect("s contains invalid character for path
-		// representation",)
-	}
-
-	#[allow(refining_impl_trait)]
-	fn as_case_convert(&self,) -> Option<&Self,>
-	{
-		Some(self,)
-	}
-}
-
-#[cfg(test)]
-mod tests
-{
-	use {
-		poison_girl_dev_error::X,
-		poison_girl_dev_test::{PoisonGirlTestB, success},
-	};
-
-	use super::*;
-
-	// Test String case detection
-	#[test]
-	fn test_string_is_camel()
-	{
-		assert!("CamelCase".to_string().is_camel());
-		assert!("PascalCase".to_string().is_camel());
-		assert!("A".to_string().is_camel());
-		assert!("ABC".to_string().is_camel());
-
-		assert!(!"camelCase".to_string().is_camel());
-		assert!(!"snake_case".to_string().is_camel());
-		assert!(!"kebab-case".to_string().is_camel());
-		assert!(!"lowercase".to_string().is_camel());
-		assert!(!"".to_string().is_camel());
-	}
-
-	#[test]
-	fn test_string_is_snake()
-	{
-		assert!("snake_case".to_string().is_snake());
-		assert!("lower_snake_case".to_string().is_snake());
-		assert!("a_b_c".to_string().is_snake());
-		assert!("single".to_string().is_snake());
-
-		assert!(!"SCREAMING_SNAKE_CASE".to_string().is_snake());
-		assert!(!"CamelCase".to_string().is_snake());
-		assert!(!"kebab-case".to_string().is_snake());
-		assert!(!"Mixed_Case".to_string().is_snake());
-	}
-
-	#[test]
-	fn test_string_is_screaming_snake()
-	{
-		assert!("SCREAMING_SNAKE_CASE".to_string().is_screaming_snake());
-		assert!("UPPER_CASE".to_string().is_screaming_snake());
-		assert!("A_B_C".to_string().is_screaming_snake());
-		assert!("SINGLE".to_string().is_screaming_snake());
-
-		assert!(!"snake_case".to_string().is_screaming_snake());
-		assert!(!"CamelCase".to_string().is_screaming_snake());
-		assert!(!"kebab-case".to_string().is_screaming_snake());
-		assert!(!"Mixed_Case".to_string().is_screaming_snake());
-	}
-
-	#[test]
-	fn test_string_is_kebab()
-	{
-		assert!("kebab-case".to_string().is_kebab());
-		assert!("lower-kebab-case".to_string().is_kebab());
-		assert!("a-b-c".to_string().is_kebab());
-		assert!("single".to_string().is_kebab());
-
-		assert!(!"UPPER-CASE".to_string().is_kebab());
-		assert!(!"CamelCase".to_string().is_kebab());
-		assert!(!"snake_case".to_string().is_kebab());
-		assert!(!"Mixed-Case".to_string().is_kebab());
-	}
-
-	// Test String case conversion
-	#[test]
-	fn test_string_to_camel()
-	{
-		let snake_case = "hello_world_test".to_string();
-		let camel: String = snake_case.to_camel();
-		// The case_transit method preserves the original spacer
-		assert_eq!(camel, "HelloWorldTest");
-
-		let kebab_case = "hello-world-test".to_string();
-		let camel: String = kebab_case.to_camel();
-		assert_eq!(camel, "HelloWorldTest");
-
-		let single_word = "hello".to_string();
-		let camel: String = single_word.to_camel();
-		assert_eq!(camel, "Hello");
-	}
-
-	#[test]
-	fn test_string_to_snake()
-	{
-		let camel_case = "HelloWorldTest".to_string();
-		let snake: String = camel_case.to_snake();
-		// The actual behavior produces "helloworldtest" because camel case
-		// doesn't have a spacer
-		assert_eq!(snake, "hello_world_test");
-
-		let kebab_case = "hello-world-test".to_string();
-		let snake: String = kebab_case.to_snake();
-		assert_eq!(snake, "hello_world_test"); // Uses original spacer
-
-		let screaming_snake = "HELLO_WORLD_TEST".to_string();
-		let snake: String = screaming_snake.to_snake();
-		assert_eq!(snake, "hello_world_test");
-	}
-
-	#[test]
-	fn test_string_to_screaming_snake()
-	{
-		let snake_case = "hello_world_test".to_string();
-		let screaming: String = snake_case.to_screaming_snake();
-		assert_eq!(screaming, "HELLO_WORLD_TEST");
-
-		let kebab_case = "hello-world-test".to_string();
-		let screaming: String = kebab_case.to_screaming_snake();
-		assert_eq!(screaming, "HELLO_WORLD_TEST"); // Uses original spacer
-
-		let camel_case = "HelloWorldTest".to_string();
-		let screaming: String = camel_case.to_screaming_snake();
-		// Due to bugs in words() method for camel case, this won't work as
-		// expected
-		assert!(!screaming.is_empty());
-	}
-
-	#[test]
-	fn test_string_to_kebab()
-	{
-		let snake_case = "hello_world_test".to_string();
-		let kebab: String = snake_case.to_kebab();
-		assert_eq!(kebab, "hello-world-test"); // Uses original spacer
-
-		let screaming_snake = "HELLO_WORLD_TEST".to_string();
-		let words = screaming_snake.words();
-		assert_eq!(words, &["HELLO", "WORLD", "TEST"]);
-		let kebab: String = screaming_snake.to_kebab();
-		assert_eq!(kebab, "hello-world-test");
-
-		let camel_case = "HelloWorldTest".to_string();
-		let kebab: String = camel_case.to_kebab();
-		// The actual behavior produces "helloworldtest" because camel case
-		// doesn't have a spacer
-		assert_eq!(kebab, "hello-world-test");
-	}
-
-	// Test find_spacer functionality
-	#[test]
-	fn test_string_find_spacer()
-	{
-		let snake_case = "hello_world".to_string();
-		let spacer: Option<String,> = snake_case.find_spacer();
-		assert_eq!(spacer, Some("_".to_string()));
-
-		let kebab_case = "hello-world".to_string();
-		let spacer: Option<String,> = kebab_case.find_spacer();
-		assert_eq!(spacer, Some("-".to_string()));
-
-		let camel_case = "HelloWorld".to_string();
-		let spacer: Option<String,> = camel_case.find_spacer();
-		assert_eq!(spacer, None);
-
-		let no_spacer = "hello".to_string();
-		let spacer: Option<String,> = no_spacer.find_spacer();
-		assert_eq!(spacer, None);
-	}
-
-	// Test words extraction
-	#[test]
-	fn test_string_words()
-	{
-		let snake_case = "hello_world_test".to_string();
-		let words = snake_case.words();
-		assert_eq!(words, vec!["hello", "world", "test"]);
-
-		let kebab_case = "hello-world-test".to_string();
-		let words = kebab_case.words();
-		assert_eq!(words, vec!["hello", "world", "test"]);
-
-		let single_word = "hello".to_string();
-		let words = single_word.words();
-		assert_eq!(words, vec!["hello"]);
-
-		// Test camel case word extraction - the current implementation has bugs
-		let camel_case = "HelloWorldTest".to_string();
-		let words = camel_case.words();
-		// The current implementation produces empty strings and doesn't work
-		// correctly Just verify it returns something
-		assert!(!words.is_empty());
-	}
-
-	// Test StringKind trait implementation for String
-	#[test]
-	fn test_string_kind_for_string()
-	{
-		let s = "test_string".to_string();
-		assert_eq!(s.dump_string(), "test_string");
-
-		let from_str: String = <String as StringKind>::from("hello",);
-		assert_eq!(from_str, "hello");
-
-		assert!(s.as_case_convert().is_some());
-	}
-
-	// Test edge cases
-	#[test]
-	fn test_empty_string_cases()
-	{
-		let empty = "".to_string();
-		assert!(!empty.is_camel());
-		assert!(empty.is_snake()); // Empty string passes snake case check
-		assert!(empty.is_screaming_snake()); // Empty string passes screaming snake check
-		assert!(empty.is_kebab()); // Empty string passes kebab case check
-
-		let spacer: Option<String,> = empty.find_spacer();
-		assert_eq!(spacer, None);
-
-		let words = empty.words();
-		assert_eq!(words, vec![""]);
-	}
-
-	#[test]
-	fn test_single_character_cases()
-	{
-		let upper = "A".to_string();
-		assert!(upper.is_camel());
-		assert!(!upper.is_snake()); // Single uppercase fails snake case
-		assert!(upper.is_screaming_snake()); // Single uppercase passes screaming snake
-		assert!(!upper.is_kebab()); // Single uppercase fails kebab case
-
-		let lower = "a".to_string();
-		assert!(!lower.is_camel());
-		assert!(lower.is_snake()); // Single lowercase passes snake case
-		assert!(!lower.is_screaming_snake()); // Single lowercase fails screaming snake
-		assert!(lower.is_kebab()); // Single lowercase passes kebab case
-	}
-
-	#[test]
-	fn test_mixed_separators()
-	{
-		let mixed = "hello_world-test".to_string();
-		// Should find the first separator it encounters
-		let spacer: Option<String,> = mixed.find_spacer();
-		assert_eq!(spacer, Some("_".to_string()));
-	}
-
-	// Test is_xxx_format_with_case helper function
-	#[test]
-	fn test_is_xxx_format_with_case()
-	{
-		// Test snake case (lowercase with underscores)
-		assert!(is_xxx_format_with_case("hello_world", Some('_'), Form::Lower));
-		assert!(!is_xxx_format_with_case(
-			"Hello_World",
-			Some('_'),
-			Form::Lower
-		));
-		assert!(!is_xxx_format_with_case(
-			"hello-world",
-			Some('_'),
-			Form::Lower
-		));
-
-		// Test screaming snake case (uppercase with underscores)
-		assert!(is_xxx_format_with_case("HELLO_WORLD", Some('_'), Form::Upper));
-		assert!(!is_xxx_format_with_case(
-			"hello_world",
-			Some('_'),
-			Form::Upper
-		));
-		assert!(!is_xxx_format_with_case(
-			"HELLO-WORLD",
-			Some('_'),
-			Form::Upper
-		));
-
-		// Test kebab case (lowercase with hyphens)
-		assert!(is_xxx_format_with_case("hello-world", Some('-'), Form::Lower));
-		assert!(!is_xxx_format_with_case(
-			"Hello-World",
-			Some('-'),
-			Form::Lower
-		));
-		assert!(!is_xxx_format_with_case(
-			"hello_world",
-			Some('-'),
-			Form::Lower
-		));
-
-		// Test camel case (starts with uppercase, no separators)
-		assert!(is_xxx_format_with_case(
-			"HelloWorld",
-			None,
-			Form::StartWithUpper
-		));
-		assert!(is_xxx_format_with_case(
-			"CamelCase",
-			None,
-			Form::StartWithUpper
-		));
-		assert!(!is_xxx_format_with_case(
-			"camelCase",
-			None,
-			Form::StartWithUpper
-		));
-		assert!(!is_xxx_format_with_case(
-			"hello_world",
-			None,
-			Form::StartWithUpper
-		));
-	}
-
-	// Test PathBuf implementations
-	#[test]
-	fn test_pathbuf_case_detection()
-	{
-		// Test with file extension - dots break case detection
-		let snake_path =
-			<std::path::PathBuf as std::convert::From<&str,>>::from(
-				"/path/to/snake_case_file.txt",
-			);
-		assert!(snake_path.is_snake());
-		assert!(!snake_path.is_camel()); // The filename doesn't start with uppercase
-
-		// Test without extension
-		let snake_no_ext =
-			<std::path::PathBuf as std::convert::From<&str,>>::from(
-				"/path/to/snake_case_file",
-			);
-		assert!(snake_no_ext.is_snake()); // snake_case_file is valid snake case
-
-		let camel_path =
-			<std::path::PathBuf as std::convert::From<&str,>>::from(
-				"/path/to/CamelCaseFile.txt",
-			);
-		assert!(camel_path.is_camel()); // CamelCaseFile.txt starts with uppercase
-		// PathBuf's is_kebab incorrectly calls is_screaming_snake, so this will
-		// be false
-		assert!(!camel_path.is_kebab()); // Note: this is due to the bug in PathBuf::is_kebab
-	}
-
-	#[test]
-	fn test_pathbuf_dump_string() -> PoisonGirlB<(),>
-	{
-		let path = <std::path::PathBuf as std::convert::From<&str,>>::from(
-			"/path/to/test_file.txt",
-		);
-		assert_eq!(path.dump_string()?, "test_file");
-
-		let path = <std::path::PathBuf as std::convert::From<&str,>>::from(
-			"simple_file.txt",
-		);
-		assert_eq!(path.dump_string()?, "simple_file");
-		X((),)
-	}
-
-	#[test]
-	fn test_pathbuf_find_spacer()
-	{
-		let snake_path =
-			<std::path::PathBuf as std::convert::From<&str,>>::from(
-				"/path/to/snake_case_file.txt",
-			);
-		let spacer: Option<String,> = snake_path.find_spacer();
-		assert_eq!(spacer, Some("_".to_string()));
-
-		let kebab_path =
-			<std::path::PathBuf as std::convert::From<&str,>>::from(
-				"/path/to/kebab-case-file.txt",
-			);
-		let spacer: Option<String,> = kebab_path.find_spacer();
-		assert_eq!(spacer, Some("-".to_string()));
-	}
-
-	#[test]
-	fn test_pathbuf_words()
-	{
-		let snake_path =
-			<std::path::PathBuf as std::convert::From<&str,>>::from(
-				"/path/to/snake_case_file.txt",
-			);
-		let words = snake_path.words();
-		assert_eq!(words, vec!["snake", "case", "file"]);
-	}
-
-	#[test]
-	fn test_pathbuf_as_string_kind()
-	{
-		let path = <std::path::PathBuf as std::convert::From<&str,>>::from(
-			"/path/to/test.txt",
-		);
-		assert!(path.as_string_kind().is_some());
-	}
-
-	#[test]
-	fn test_pathbuf_as_case_convert()
-	{
-		let path = <std::path::PathBuf as std::convert::From<&str,>>::from(
-			"/path/to/test.txt",
-		);
-		assert!(path.as_case_convert().is_some());
-	}
-
-	#[test]
-	#[should_panic(expected = "you should not use `PathBuf::from`")]
-	fn test_pathbuf_from_unimplemented()
-	{
-		let _: PathBuf = <PathBuf as StringKind>::from("test",);
-	}
-
-	// Test trait implementations
-	#[test]
-	fn test_str_enhanced_trait()
-	{
-		let s = "test_string".to_string();
-		// StrEnhanced is a marker trait that combines CaseConvert + StringKind
-		// We can test that it's implemented by using its methods
-		assert!(s.is_snake()); // test_string is valid snake case
-		assert_eq!(s.dump_string(), "test_string");
-	}
-
-	// Property-based tests
-	// quickcheck! {
-	// 	fn test_case_conversion_roundtrip_snake_to_snake(s :String)->bool {
-	// 		let original = s.clone();
-	// 		let converted: String = original.to_snake();
-	// 		// Converting snake case to snake case should be idempotent
-	// 		converted.to_lowercase().replace("_", "_")== s.to_lowercase()
-	// 	}
-	//
-	// 	fn test_case_conversion_preserves_length_concept(s in
-	// "[a-zA-Z][a-zA-Z0-9_-]{2,}") { 		let original = s.clone();
-	// 		let _words_count = original.words().len();
-	//
-	// 		// Converting to different cases should preserve the concept of word
-	// count 		// Only test if the string has enough length to avoid the
-	// slicing bug 		if original.len() >= 2 && !original.is_empty() {
-	// 			// Test that conversions don't panic and produce non-empty results
-	// 			let snake_result = std::panic::catch_unwind(|| {
-	// 				let snake: String = original.to_snake();
-	// 				snake
-	// 			});
-	//
-	// 			let camel_result = std::panic::catch_unwind(|| {
-	// 				let camel: String = original.to_camel();
-	// 				camel
-	// 			});
-	//
-	// 			let screaming_result = std::panic::catch_unwind(|| {
-	// 				let screaming: String = original.to_screaming_snake();
-	// 				screaming
-	// 			});
-	//
-	// 			let kebab_result = std::panic::catch_unwind(|| {
-	// 				let kebab: String = original.to_kebab();
-	// 				kebab
-	// 			});
-	//
-	// 			// All conversions should either succeed or fail gracefully
-	// 			if let Ok(snake) = snake_result {
-	// 				prop_assert!(!snake.is_empty());
-	// 			}
-	// 			if let Ok(camel) = camel_result {
-	// 				prop_assert!(!camel.is_empty());
-	// 			}
-	// 			if let Ok(screaming) = screaming_result {
-	// 				prop_assert!(!screaming.is_empty());
-	// 			}
-	// 			if let Ok(kebab) = kebab_result {
-	// 				prop_assert!(!kebab.is_empty());
-	// 			}
-	// 		}
-	// 	}
-	//
-	// 	fn test_spacer_detection_consistency(s in "[a-z]+[_-][a-z]+") {
-	// 		let spacer: Option<String> = s.find_spacer();
-	// 		prop_assert!(spacer.is_some());
-	// 		let spacer = spacer.unwrap();
-	// 		prop_assert!(spacer == "_" || spacer == "-");
-	// 		prop_assert!(s.contains(&spacer));
-	// 	}
-	//
-	// 	fn test_case_detection_mutual_exclusivity_snake_camel(s in
-	// "[a-zA-Z][a-zA-Z0-9_]*") { 		// Test that strings with underscores and
-	// mixed case behave as expected 		if s.contains("_") && s.chars().any(|c|
-	// c.is_uppercase()) { 			// Mixed case with underscores should not be
-	// valid snake case 			prop_assert!(!s.is_snake());
-	// 		} else if s.contains("_") && s.chars().all(|c| c.is_ascii_lowercase()
-	// || c == '_') { 			// Pure lowercase letters with underscores should be
-	// valid snake case 			// Note: numeric characters are NOT allowed in snake
-	// case by this implementation 			prop_assert!(s.is_snake());
-	// 		}
-	// 	}
-	//
-	// 	fn test_pathbuf_filename_extraction(filename in
-	// "[a-zA-Z][a-zA-Z0-9_.-]*") { 		let path = <std::path::PathBuf as
-	// std::convert::From<String>>::from(format!("/some/path/{}", filename));
-	// 		prop_assert_eq!(path.dump_string(),
-	// filename.split('.').nth(0).unwrap()); 	}
-	// }
-
-	// Integration tests combining multiple traits
-	#[test]
-	fn test_string_enhanced_integration()
-	{
-		let snake_str = "hello_world_test".to_string();
-
-		// Test that we can use both CaseConvert and StringKind methods
-		assert!(snake_str.is_snake()); // hello_world_test is valid snake case
-		assert_eq!(snake_str.dump_string(), "hello_world_test");
-
-		// Test case conversion
-		let camel: String = snake_str.to_camel();
-		// The case_transit method preserves the original spacer
-		assert_eq!(camel, "HelloWorldTest");
-		assert!(camel.is_camel());
-
-		// Test that converted string also implements the traits
-		let back_to_snake: String = camel.to_snake();
-		// Due to bugs in words() method, this won't work as expected
-		assert!(!back_to_snake.is_empty());
-	}
-
-	#[test]
-	fn test_pathbuf_enhanced_integration() -> PoisonGirlTestB
-	{
-		let snake_path =
-			<std::path::PathBuf as std::convert::From<&str,>>::from(
-				"/path/to/hello_world_test.txt",
-			);
-
-		assert!(snake_path.is_snake());
-		assert_eq!(snake_path.dump_string()?, "hello_world_test");
-
-		// Test spacer detection
-		let spacer: Option<String,> = snake_path.find_spacer();
-		assert_eq!(spacer, Some("_".to_string()));
-
-		// Test word extraction
-		let words = snake_path.words();
-		assert_eq!(words, vec!["hello", "world", "test"]);
-		success!()
-	}
-
-	// Error condition tests
-	#[test]
-	#[should_panic(expected = "failed to get file/dir name")]
-	fn test_pathbuf_dump_string_no_filename()
-	{
-		let path =
-			<std::path::PathBuf as std::convert::From<&str,>>::from("/",);
-		let _ = path.dump_string();
-	}
-
-	// Test special characters and edge cases
-	#[test]
-	fn test_special_characters_in_strings()
-	{
-		let with_numbers = "test123_case456".to_string();
-		// Numbers are allowed in snake case according to the implementation
-		assert!(with_numbers.is_snake());
-
-		let with_numbers_camel = "Test123Case456".to_string();
-		assert!(with_numbers_camel.is_camel());
-	}
-
-	#[test]
-	fn test_case_conversion_with_numbers()
-	{
-		let snake_with_nums = "hello123_world456".to_string();
-		let camel: String = snake_with_nums.to_camel();
-		// The case_transit method preserves the original spacer
-		assert_eq!(camel, "Hello123World456");
-
-		let screaming: String = snake_with_nums.to_screaming_snake();
-		assert_eq!(screaming, "HELLO123_WORLD456");
-	}
-
-	// Test boundary conditions
-	#[test]
-	fn test_very_long_strings()
-	{
-		let long_snake = "a".repeat(100,) + "_" + &"b".repeat(100,);
-		assert!(long_snake.is_snake()); // Long snake case string is valid
-
-		let words = long_snake.words();
-		assert_eq!(words.len(), 2);
-		assert_eq!(words[0].len(), 100);
-		assert_eq!(words[1].len(), 100);
-	}
-
-	#[test]
-	fn test_unicode_handling()
-	{
-		let unicode_snake = "hello_世界".to_string();
-		assert!(!unicode_snake.is_snake());
-		assert!(!unicode_snake.is_screaming_snake());
-		assert_eq!(
-			unicode_snake.find_spacer::<String>(),
-			Some("_".to_string())
-		);
-		assert_eq!(unicode_snake.words(), vec!["hello", "世界"]);
-
-		let converted: String = unicode_snake.to_snake();
-		assert_eq!(converted, "hello_世界");
-
-		let unicode_word = "Καλημέρα".to_string();
-		assert!(!unicode_word.is_camel());
-		assert!(!unicode_word.is_snake());
-		assert!(!unicode_word.is_screaming_snake());
-		assert!(!unicode_word.is_kebab());
-		assert_eq!(unicode_word.words(), vec!["Καλημέρα"]);
-	}
-
-	#[test]
-	fn test_extreme_edge_cases()
-	{
-		// Test with extreme edge cases
-		let single_char = "a".to_string();
-		assert!(!single_char.is_camel()); // Single lowercase is not camel
-		assert!(single_char.is_snake()); // Single lowercase is valid snake
-		assert!(!single_char.is_screaming_snake()); // Single lowercase is not screaming
-		assert!(single_char.is_kebab()); // Single lowercase is valid kebab
-
-		let single_upper = "A".to_string();
-		assert!(single_upper.is_camel()); // Single uppercase is camel
-		assert!(!single_upper.is_snake()); // Single uppercase is not snake
-		assert!(single_upper.is_screaming_snake()); // Single uppercase is screaming
-		assert!(!single_upper.is_kebab()); // Single uppercase is not kebab
-	}
-
-	#[test]
-	fn test_numeric_only_strings()
-	{
-		// Test with numeric-only strings
-		let numbers = "123456".to_string();
-		assert!(!numbers.is_camel()); // Numbers don't start with uppercase
-		assert!(numbers.is_snake()); // Numbers are valid in snake case
-		assert!(numbers.is_screaming_snake()); // Numbers are valid in screaming snake
-		assert!(numbers.is_kebab()); // Numbers are valid in kebab case
-
-		let mixed_numbers = "123_456".to_string();
-		assert!(mixed_numbers.is_snake()); // Numbers with underscores
-		assert!(!mixed_numbers.is_camel()); // Doesn't start with uppercase
-	}
-
-	#[test]
-	fn test_special_character_combinations()
-	{
-		// Test various special character combinations
-		let underscore_only = "___".to_string();
-		assert!(underscore_only.is_snake()); // Only underscores is actually valid snake case
-		assert!(!underscore_only.is_camel()); // Only underscores is not camel
-
-		let dash_only = "---".to_string();
-		assert!(dash_only.is_kebab()); // Only dashes is actually valid kebab case
-		assert!(!dash_only.is_camel()); // Only dashes is not camel
-
-		let mixed_separators = "hello_world-test".to_string();
-		// Should find the first separator (underscore)
-		let spacer: Option<String,> = mixed_separators.find_spacer();
-		assert_eq!(spacer, Some("_".to_string()));
-	}
-
-	#[test]
-	fn test_case_conversion_edge_cases()
-	{
-		// Test case conversion with edge cases
-		let empty = "".to_string();
-		let words = empty.words();
-		assert_eq!(words, vec![""]);
-
-		// Test with single character
-		let single = "a".to_string();
-		let camel: String = single.to_camel();
-		assert_eq!(camel, "A");
-
-		let screaming: String = single.to_screaming_snake();
-		assert_eq!(screaming, "A");
-	}
-
-	#[test]
-	fn test_words_method_edge_cases()
-	{
-		// Test the words method with various edge cases
-		let consecutive_separators = "hello__world".to_string();
-		let words = consecutive_separators.words();
-		assert_eq!(words, vec!["hello", "", "world"]);
-
-		let leading_separator = "_hello_world".to_string();
-		let words = leading_separator.words();
-		assert_eq!(words, vec!["", "hello", "world"]);
-
-		let trailing_separator = "hello_world_".to_string();
-		let words = trailing_separator.words();
-		assert_eq!(words, vec!["hello", "world", ""]);
-	}
-
-	#[test]
-	fn test_camel_case_variations()
-	{
-		// Test various camel case patterns
-		let pascal_case = "PascalCase".to_string();
-		assert!(pascal_case.is_camel());
-
-		let camel_case = "camelCase".to_string();
-		assert!(!camel_case.is_camel()); // Doesn't start with uppercase
-
-		let all_caps = "ALLCAPS".to_string();
-		assert!(all_caps.is_camel()); // Starts with uppercase
-
-		let mixed_case = "MiXeDcAsE".to_string();
-		assert!(mixed_case.is_camel()); // Starts with uppercase
-	}
-
-	#[test]
-	fn test_snake_case_variations()
-	{
-		// Test various snake case patterns
-		let standard_snake = "hello_world".to_string();
-		assert!(standard_snake.is_snake());
-
-		let snake_with_numbers = "hello_123_world".to_string();
-		assert!(snake_with_numbers.is_snake());
-
-		let leading_underscore = "_hello_world".to_string();
-		assert!(leading_underscore.is_snake());
-
-		let trailing_underscore = "hello_world_".to_string();
-		assert!(trailing_underscore.is_snake());
-
-		let multiple_underscores = "hello__world".to_string();
-		assert!(multiple_underscores.is_snake());
-	}
-
-	#[test]
-	fn test_screaming_snake_case_variations()
-	{
-		// Test various screaming snake case patterns
-		let standard_screaming = "HELLO_WORLD".to_string();
-		assert!(standard_screaming.is_screaming_snake());
-
-		let screaming_with_numbers = "HELLO_123_WORLD".to_string();
-		assert!(screaming_with_numbers.is_screaming_snake());
-
-		let leading_underscore = "_HELLO_WORLD".to_string();
-		assert!(leading_underscore.is_screaming_snake());
-
-		let trailing_underscore = "HELLO_WORLD_".to_string();
-		assert!(trailing_underscore.is_screaming_snake());
-	}
-
-	#[test]
-	fn test_kebab_case_variations()
-	{
-		// Test various kebab case patterns
-		let standard_kebab = "hello-world".to_string();
-		assert!(standard_kebab.is_kebab());
-
-		let kebab_with_numbers = "hello-123-world".to_string();
-		assert!(kebab_with_numbers.is_kebab());
-
-		let leading_dash = "-hello-world".to_string();
-		assert!(leading_dash.is_kebab());
-
-		let trailing_dash = "hello-world-".to_string();
-		assert!(trailing_dash.is_kebab());
-
-		let multiple_dashes = "hello--world".to_string();
-		assert!(multiple_dashes.is_kebab());
-	}
-
-	#[test]
-	fn test_case_transit_method()
-	{
-		// Test the case_transit method directly
-		let snake_case = "hello_world".to_string();
-		let upper_converter = |s: String| s.to_uppercase();
-		let result: String =
-			snake_case.case_transit(upper_converter, Some('_',),);
-		assert_eq!(result, "HELLO_WORLD");
-
-		let lower_converter = |s: String| s.to_lowercase();
-		let result: String =
-			snake_case.case_transit(lower_converter, Some('_',),);
-		assert_eq!(result, "hello_world");
-	}
-
-	#[test]
-	fn test_is_xxx_format_with_case_edge_cases()
-	{
-		// Test the helper function with edge cases
-		assert!(!is_xxx_format_with_case("", None, Form::StartWithUpper)); // Empty string doesn't start with uppercase
-		assert!(is_xxx_format_with_case("", Some('_'), Form::Lower)); // Empty string with separator
-		assert!(is_xxx_format_with_case("", Some('-'), Form::Upper)); // Empty string with separator
-
-		// Test with single characters
-		assert!(is_xxx_format_with_case("A", None, Form::StartWithUpper));
-		assert!(!is_xxx_format_with_case("a", None, Form::StartWithUpper));
-		assert!(is_xxx_format_with_case("a", Some('_'), Form::Lower));
-		assert!(is_xxx_format_with_case("A", Some('_'), Form::Upper));
-	}
-
-	#[test]
-	fn test_pathbuf_edge_cases()
-	{
-		// Test PathBuf implementations with edge cases
-		let root_path =
-			<std::path::PathBuf as std::convert::From<&str,>>::from("/",);
-		// This should panic when trying to get filename, but we'll catch it
-		let result = std::panic::catch_unwind(|| root_path.dump_string(),);
-		assert!(result.is_err());
-
-		// Test with current directory
-		let current_dir =
-			<std::path::PathBuf as std::convert::From<&str,>>::from(".",);
-		let result = std::panic::catch_unwind(|| current_dir.dump_string(),);
-		// This might succeed or fail depending on the system
-		let _ = result;
-	}
-
-	#[test]
-	fn test_pathbuf_case_detection_edge_cases()
-	{
-		// Test PathBuf case detection with various filename patterns
-		let snake_file =
-			<std::path::PathBuf as std::convert::From<&str,>>::from(
-				"/path/to/snake_case_file",
-			);
-		assert!(snake_file.is_snake());
-
-		let camel_file =
-			<std::path::PathBuf as std::convert::From<&str,>>::from(
-				"/path/to/CamelCaseFile",
-			);
-		assert!(camel_file.is_camel());
-
-		let kebab_file =
-			<std::path::PathBuf as std::convert::From<&str,>>::from(
-				"/path/to/kebab-case-file",
-			);
-		assert!(kebab_file.is_kebab());
-
-		let screaming_file =
-			<std::path::PathBuf as std::convert::From<&str,>>::from(
-				"/path/to/SCREAMING_SNAKE_FILE",
-			);
-		assert!(screaming_file.is_screaming_snake());
-	}
-
-	#[test]
-	fn test_pathbuf_with_extensions() -> PoisonGirlTestB
-	{
-		// Test PathBuf behavior with various file extensions
-		let extensions = vec![
-			".txt", ".rs", ".toml", ".md", ".json", ".xml", ".html", ".css",
-			".js",
-		];
-
-		for ext in extensions {
-			let filename = format!("test_file{}", ext);
-			let path =
-				<std::path::PathBuf as std::convert::From<String,>>::from(
-					format!("/path/to/{}", filename),
-				);
-			assert_eq!(path.dump_string()?, filename.split('.').next()?);
-
-			// Test case detection with extensions
-			let snake_with_ext = format!("snake_case_file{}", ext);
-			let path =
-				<std::path::PathBuf as std::convert::From<String,>>::from(
-					format!("/path/to/{}", snake_with_ext),
-				);
-			assert!(path.is_snake());
-		}
-		success!()
-	}
-
-	#[test]
-	fn test_memory_efficiency()
-	{
-		// Test that operations don't cause excessive memory usage
-		let large_string = "a".repeat(10000,) + "_" + &"b".repeat(10000,);
-
-		// These operations should complete without excessive memory usage
-		let is_snake = large_string.is_snake();
-		assert!(is_snake);
-
-		let words = large_string.words();
-		assert_eq!(words.len(), 2);
-		assert_eq!(words[0].len(), 10000);
-		assert_eq!(words[1].len(), 10000);
-
-		// Test case conversion with large strings
-		let camel: String = large_string.to_camel();
-		assert!(!camel.is_empty());
-	}
-
-	#[test]
-	fn test_string_kind_trait_completeness()
-	{
-		// Test StringKind trait implementation completeness
-		let s = "test_string".to_string();
-
-		// Test all trait methods
-		let dumped = s.dump_string();
-		assert_eq!(dumped, "test_string");
-
-		let from_str: String = <String as StringKind>::from("hello",);
-		assert_eq!(from_str, "hello");
-
-		let case_convert = s.as_case_convert();
-		assert!(case_convert.is_some());
-
-		let string_kind = s.as_string_kind();
-		assert!(string_kind.is_some());
-	}
-
-	#[test]
-	fn test_pathbuf_string_kind_trait() -> PoisonGirlTestB
-	{
-		// Test StringKind trait implementation for PathBuf
-		let path = <std::path::PathBuf as std::convert::From<&str,>>::from(
-			"/path/to/test.txt",
-		);
-
-		let dumped = path.dump_string()?;
-		assert_eq!(dumped, "test");
-
-		let case_convert = path.as_case_convert();
-		assert!(case_convert.is_some());
-
-		let string_kind = path.as_string_kind();
-		assert!(string_kind.is_some());
-		success!()
-	}
-}
+pub mod case_conversion;
+pub mod toml_tools;
