+++ crates/error/dev/Cargo.toml
@@ -8,6 +8,7 @@ ovmf-prebuilt = { workspace = true }
 poison_girl_this_is_b_wrapper_dev = { workspace = true }
 toml = { workspace = true }
 hadris-fat = { workspace = true }
+strum = { workspace = true }
 
 [lints]
 workspace = true
